python/requirements.txt
argcomplete==1.9.3
base58==1.0.3
boto3==1.9.197
botocore==1.12.197
certifi==2019.6.16
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
colander==1.7.0
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-environ==0.4.5
django-extensions==2.2.1
django-storages==1.7.1
docutils==0.15.1
durationpy==0.5
ecdsa==0.13.2
future==0.16.0
hjson==3.0.1
idna==2.8
iso8601==0.1.12
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
libeospy==2.0.1
oauthlib==3.0.2
placebo==0.9.0
psycopg2-binary==2.8.3
python-dateutil==2.6.1
python-slugify==1.2.4
python-twitter==3.5
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
translationstring==1.3
troposphere==2.4.9
Unidecode==1.1.1
urllib3==1.25.3
Werkzeug==0.15.5
wsgi-request-logger==0.4.6
zappa==0.48.2
